app/functions/signals/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
numpy==1.26.4
scikit-learn==1.4.2